Moves mesh creation functions in a new file
Adds a new file, tdycoremesh.c, that handles creation of the mesh
